Fixed problems with the documentation
